pyproject.toml
[project]
name = "vue-modorn-commercial-web-nchuit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
    "google-generativeai>=0.8.5",
    "pymysql>=1.1.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "types-flask-cors>=5.0.0.20250413",
    "types-pymysql>=1.1.0.20241103",
    "types-requests>=2.32.0.20250328",
]
